Rename statistics module
