+++ magali/_stats.py
